+++ src/Behat/Context/NavigationContext.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace Behat\Context;
+
+use Behat\Behat\Context\SnippetAcceptingContext;
+use Behat\Behat\Hook\Scope\BeforeScenarioScope;
+use Behat\Gherkin\Node\TableNode;
+use Behat\Mink\Exception\ExpectationException;
+use Behat\MinkExtension\Context\MinkContext;
+use Behat\Symfony2Extension\Context\KernelAwareContext;
+use Behat\Symfony2Extension\Context\KernelDictionary;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Matcher\UrlMatcher;
+
+/**
+ * @author    Léo Benoist <leo.benoist@akeneo.com>
+ * @copyright 2016 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ */
+class NavigationContext implements SnippetAcceptingContext, KernelAwareContext
+{
+    use KernelDictionary;
+
+    /** @var MinkContext */
+    protected $minkContext;
+    
+    /**
+     * @BeforeScenario
+     *
+     * @param BeforeScenarioScope $scope
+     */
+    public function gatherContexts(BeforeScenarioScope $scope)
+    {
+        $environment = $scope->getEnvironment();
+
+        $this->minkContext = $environment->getContext('Behat\MinkExtension\Context\MinkContext');
+    }
+
+    /**
+     * @param string $page
+     *
+     * Checks, that current page PATH is equal to specified.
+     *
+     * @Then I should be redirected to :page
+     */
+    public function assertPageAddress($page)
+    {
+        $this->minkContext->assertPageAddress($page);
+    }
+}
